graphic-interface/requirements.txt
biopython
matplotlib
pandas
python-dotenv
tensorflow
regex
seaborn

Flask==2.2.5
Flask-Celery==2.4.3
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-Script==2.0.6
